backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
pydantic
email-validator
pandas
openpyxl
python-multipart
